// Frontend/src/models/Verb.ts
export default interface Verb {
    id: number;
    infinitive: string;
    praesens_ich: string;
    praesens_du: string;
    praesens_er: string;
    praeteritum_ich: string;
    partizip_II: string;
    konjunktiv_II_ich: string;
    imperativ_singular: string;
    imperativ_plural: string;
    hilfsverb: string;
}
  